(story) move subcomponent stories to subfolder, remove props from SaturationFilter story
