+++ frontend/src/components/SaturationFilter/stories/SaturationFilter.stories.tsx
@@ -1,14 +1,14 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import { RecoilRoot } from "recoil";
-import SaturationFilter from "./SaturationFilter";
+import SaturationFilter from "../SaturationFilter";
 
 export default {
 	component: SaturationFilter,
 	argTypes: {},
 } as ComponentMeta<typeof SaturationFilter>;
 
-export const Basic: ComponentStory<typeof SaturationFilter> = (props) => (
+export const Basic: ComponentStory<typeof SaturationFilter> = () => (
 	<RecoilRoot>
-		<SaturationFilter {...props} />
+		<SaturationFilter />
 	</RecoilRoot>
 );
